fix bug bullet fly, kill enemy when out of health

+++ src/main/java/com/group17/towerdefense/repositories/entity/FireableSingleEntity.java
@@ -2,5 +2,5 @@ package com.group17.towerdefense.repositories.entity;
 
 public interface FireableSingleEntity {
     public GameEntity getTarget();
-    public double getDamage();
+    public int getDamage();
 }
